WIP: Extend base foundation with variant specific foundations
