+++ packages/mdc-top-app-bar/short/foundation.js
@@ -1,0 +1,78 @@
+/**
+ * @license
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import MDCTopAppBarFoundation from '../foundation';
+
+/**
+ * @extends {MDCTopAppBarFoundation<!MDCShortTopAppBarFoundation>}
+ * @final
+ */
+class MDCShortTopAppBarFoundation extends MDCTopAppBarFoundation {
+  /**
+   * @param {!MDCTopAppBarAdapter} adapter
+   */
+  constructor(adapter) {
+    super(Object.assign(MDCTopAppBarFoundation.defaultAdapter, adapter));
+    // State variable for the current top app bar state
+    /** @private {boolean} */
+    this.isCollapsed_ = false;
+
+    this.scrollHandler_ = () => this.shortAppBarScrollHandler_();
+  }
+
+  init() {
+    super.init();
+
+    const isAlwaysCollapsed = this.adapter_.hasClass(MDCTopAppBarFoundation.cssClasses.SHORT_COLLAPSED_CLASS);
+
+    if (this.adapter_.getTotalActionItems() > 0) {
+      this.adapter_.addClass(MDCTopAppBarFoundation.cssClasses.SHORT_HAS_ACTION_ITEM_CLASS);
+    }
+
+    if (!isAlwaysCollapsed) {
+      this.adapter_.registerScrollHandler(this.scrollHandler_);
+      this.shortAppBarScrollHandler_();
+    }
+  }
+
+  destroy() {
+    super.destroy();
+    this.adapter_.deregisterScrollHandler(this.scrollHandler_);
+  }
+
+  /**
+   * Scroll handler for applying/removing the collapsed modifier class
+   * on the short top app bar.
+   */
+  shortAppBarScrollHandler_() {
+    const currentScroll = this.adapter_.getViewportScrollY();
+
+    if (currentScroll <= 0) {
+      if (this.isCollapsed_) {
+        this.adapter_.removeClass(MDCTopAppBarFoundation.cssClasses.SHORT_COLLAPSED_CLASS);
+        this.isCollapsed_ = false;
+      }
+    } else {
+      if (!this.isCollapsed_) {
+        this.adapter_.addClass(MDCTopAppBarFoundation.cssClasses.SHORT_COLLAPSED_CLASS);
+        this.isCollapsed_ = true;
+      }
+    }
+  }
+}
+
+export default MDCShortTopAppBarFoundation;
